Remove FluxEstimate
